+++ src/libuc/database/lang.txt
@@ -1,0 +1,1 @@
+/* lang=C++20 */
