+++ src/libuc/database/minhdrs.txt
@@ -1,0 +1,5 @@
+#include	<envstandards.h>	/* ordered first to configure */
+#include	<utypedefs.h>
+#include	<utypealiases.h>
+#include	<clanguage.h>
+#include	<localmisc.h>
