+++ src/libuc/database/ns.txt
@@ -1,0 +1,6 @@
+/* imported namespaces */
+
+
+/* local typedefs */
+
+
