+++ src/libuc/database/sccs.txt
@@ -1,0 +1,1 @@
+/* version %I% last-modified %G% */
